+++ NEWS.md
@@ -1,5 +1,6 @@
 # camtrapdp (development version)
 
+* New function `write_eml()` transforms Camtrap DP metadata to EML (#61).
 * New function `round_coordinates()` allows to fuzzy/generalize location information by rounding deployment `latitude` and `longitude`. It also updates `coordinateUncertainty` in the deployments and `coordinatePrecision` in the metadata (#85).
 
 # camtrapdp 0.3.1